In-mem MemStorage notifications are handled sequentially
This patch ensures that notifications sent by the in-memory storage
are handled sequentially. More specifically, it ensures that method
emitUpdates of class ClassNode is completed before being called again,
which was previously resulting in two observables being emitted for
the same topology object. This was happening when emitting the 1st
observable on variable stream would result in an update to storage,
which would in turn call emitUpdates a 2nd time. This 2nd call would
emit a 2nd observable on variable stream before the 1st call to
emitUpdates would be finished.

To prevent this undesired scenario from happening, calls to
emitUpdates are scheduled on an executor that uses the same thread as
the submitter.

Change-Id: I13d0f9cb5ca13b5bf07096b9a92837f851b688e3
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>