VPNaaS: Allow vpn service to be updated after container setup
Right now, once the ipsec container is bound to a port, its
configuration cannot be changed (e.g. associated ipsec connections).
Once the container scheduler is in, this static configuration will no
longer be usable as the binding could be possibly done before new ipsec
connections are added. This patch enables the ipsec container handler to
monitor updates on the vpnservice associated, so the ipsec namespace
can be recreated if needed.

Change-Id: I3bfa8eebe2df0149ac72906359d9671ff3db7a27
Signed-off-by: Xavi León <xavi.leon@midokura.com>