VPNaaS: monitor changes on vpn service and ipsec connections
Right now, once the ipsec container is bound to a port, its
configuration cannot be changed (e.g. vpn services and
associated ipsec connections).
This patch enables the ipsec container handler to monitor
updates on the vpnservice and ipsec site connections associated,
so the ipsec namespace can be recreated if needed.

Change-Id: Ib2f37aceba11f1cc99e887be3fc5efa83d35ea3d
Signed-off-by: Xavi León <xavi.leon@midokura.com>