VPNaaS: Introduce host container weight
The topology model, API and CLI introduce a host container weight.
This allows the cluster container service to schedule containers
according to weights, where a host with a higher weight has a
greater proababilty of hosting a container.

Container weights are integers greater or equal to zero. The
default weight is 1. A weight of zero (0) means that the host
cannot be used to schedule a container.

Change-Id: Ibc13bbe2144112688f6b5a7d20e403a821d7da5e
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>