VPNaaS: consider VPNService adminStateUp
The IPSecContainer will not spawn a namespace unless there is at least
one VpnService with adminStateUp.  In this case, the container status
will be shown as RUNNING, indicating that the agent was able to handle
the container, but the namespace will be set to an empty string. We
might chose to add a new DISABLED state to reflect this later.

Change-Id: I9a96c2a3a27b0923ab302c5ec3b2910bd708f335
Signed-off-by: Galo Navarro <galo@midokura.com>