VPNaaS: Agent container service uses host weight
This change allows the agent container service to report the
container weight configured in the topology, instead of the
default value (1). This value is reported to the cluster via
the container service status. This allows the agent to use
additional sources of configurations for the container weight
(e.g. set the default using mn-conf).

Change-Id: I0a24020d9b4f08a836c93ea56426df1b1bac42f4
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>