Router peering: Correctly support ICMP
This patch ensures we properly support ICMP in the router peering
feature. ICMP packets present two challenges:

1) Stateful ICMP echo flows are userspace flows since OVS doesn't
   allow us to match on the ICMP ID. We consider these flows userspace
   only flows and we don't install any kernel flows for them, so we
   are unable to properly recirculate the packet.
2) The same goes for ICMP errors, with the added complication that we
   actually mangle the payload of these packets when doing or
   reversing NAT.

As a consequence, we stop executing the packet after flow
installation, which allowed us to encapsulate the packet in the
kernel, and instead encapsulate the packet ourselves in usermode. This
allows us to do any mangling we need, and it also solves the userspace
flow problem.

Change-Id: I5a8f4ce28f1f17117935a9ac055bf690538f286f
Signed-off-by: Duarte Nunes <duarte@midokura.com>