ZOOM class observable recovers when closed
Currently, the ZOOM object and state observables recover when the
the underlying watchers to ZooKeeper have been closed because all
previous subscribers have unsubscribed. However, the class
observable does not provide the same behavior, and furthermore the
exception returned indicating that the observable is closed is also
used when the connection to ZooKeeper is lost.

This change adds a new exception to signal that the underlying
cache for a class observable is closed, and ensures that ZOOM
recovers the observables that emit this error.

Change-Id: I61cfbe0a1200f0a62a196b7f9301adea81bb4c0b
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>