Container management: elect leader among minions
This patch adds leader election to the ContainerService minion, so that
only one Cluster node will perform container allocation tasks at any
given point in time.  Some minor refactors to adapt the
MidonetBackendTestModule for easier injection.

We also fix a bug in error handling: the service will recover after
errors are emitted on the Scheduler observable, until a max # of errors
in which case it'll give up the leader latch and set the Minion as
failed.

TODO: review the error handling policy, we probably want to reset the
error count when the recoverd Observable is deemed healthy.

TODO: adapt the LeaderLatch used in VxlanGatewayService to follow a
similar approach, not doing it in this patch to contain scope.

Change-Id: I5917c2ffa565419e0546a2752afdccb4d28602f4
Signed-off-by: Galo Navarro <galo@midokura.com>