Container management: elect leader among minions
This patch adds leader election to the ContainerService minion, so that
only one Cluster node will perform container allocation tasks at any
given point in time.  Some minor refactors to adapt the
MidonetBackendTestModule for easier injection.

TODO: adapt the LeaderLatch used in VxlanGatewayService to follow a
similar approach, not doing it in this patch to contain scope.

Change-Id: I5917c2ffa565419e0546a2752afdccb4d28602f4
Signed-off-by: Galo Navarro <galo@midokura.com>