VPNaaS: not call init_conns on non-existing site connection config
All connections for a given vpnservice break when setting the
adminStateUp property to DOWN to only one of them. This was because the
configuration for that site connection was not written to the config but
we still tried to init it. This patch solves this problem and cleans up
some conditionals for better readability.

Change-Id: I0e9e5283fa5a06cb6b32ac37586cf79964154122
Signed-off-by: Xavi León <xavi.leon@midokura.com>