Consolidate ARP and MAC table seeding
We were not seeding ARP and MAC tables for RouterGateway and
RouterInterface ports, and were unnecessarily separating the code for
seeding them for remote site and VIF ports. This change consolidates
creation and deletion of seeded entries for all four types of ports
that need them.

Also deletes anti-spoof chain and removes IP address from IP address
group when converting a VIF port to a router interface port.

Change-Id: I954ef8cb4d149ba79ced43ece4daf614d9090b3a
Signed-off-by: Brandon Berg <bberg@midokura.com>