VPNaaS: make sure we unsubscribe from vpnservice observable
When we removed all ipsec site connections or we set their adminStateUp
to DOWN, we set the configuration to null so at the time of removing
the associated vpnservice and call the delete method, we do not
unsubscribe from the vpn service observable.

This means that upon updating the same router (e.g. by adding another
vpnservice) the former observable will emit a new notification which
will create a clone namespace for the new vpnservice using the old
container port (messing everything up).

This patch ensures that when deleting the container, we unsubscribe
from the vpn service observable even if the configuration was set
to null.

Change-Id: I9bb32d2202768c5da1f7f036735fc8340c45bfd4
Signed-off-by: Xavi León <xavi.leon@midokura.com>