VPNaaS: make sure container port/ip on router is unused
Right now, we use a fixed subnet, and ips for the port connected to the
vpn container. This patch makes sure the subnet/ip used for the
container port is not used (by looking into the ip addresses of the
associated router) by looking for a free subnet in a subnet range
reserved for containers. This way, other containers can take advantage
of this allocation mechanism and prevents clashing of unrelated containers
types (e.g. vpn container and haproxy containers once they are moved to
the container framework).

Change-Id: I3e58f87cc707b283a61726b379ad0c55d4973e5b
Signed-off-by: Xavi León <xavi.leon@midokura.com>