Update patch set 1
Patch Set 1:

(1 comment)

The subnet selection is indeed executed in the translation (the code is in nsdb so it can be reused by other container translations that may use containers in the future). Disregard the race condition comment, it's safe given that the translation is in a transaction.

Patch-set: 1
Label: Code-Review=0