Serialize flow trace logs to Cassandra
We were sending them asynchronously and this allows for race conditions
where the order of events as they get stored in Cassandra doesn't match
simulation order.

In this patch the logger will use a dedicated thread to push requests
synchronously to Cassandra.  The queue is implemented with a non
blocking bounded queue, and a Thread that consumes it.

MDTS tests require a new wait before the event counts are extracted from
cassandra, as the inserts are triggered outside of the sim. threads, so
the delay is potentially higher.

Change-Id: Id758e4ce3b378fe4f5231977bb7b1a11c2d7f08f
Signed-off-by: Galo Navarro <galo@midokura.com>