Serialize flow trace logs to Cassandra
We were sending them asynchronously and this allows for race conditions
where the order of events as they get stored in Cassandra doesn't match
simulation order.

In this patch the logger will use a dedicated thread to push requests
synchronously to Cassandra.  A simple onBackpressureDrop strategy is
used, I'd rather use a bounded buffer with a Drop on overflow but this
strategy is still not in RxJava.  I have submitted a PR to RxJava (3487)
that would allow this.

Change-Id: I2af42d61362a5a6d967541d7f270bfd668c52cca
Signed-off-by: Galo Navarro <galo@midokura.com>