Prefer nextHopGateway over ARP cache
* If there is an IPv4 packet going through a router, and the final
destination IP is in the arp cache, but the routes have directed the
packet to a next hop (and therefore decided the packet is not destined
for the local subnet), then the packet should not be delivered to the
MAC address associated with the IP in the arp cache.

* This fix causes the next hop MAC resolution code to first choose the IP,
which will either be the next hop IP or the final destination, then get the
MAC address.

Change-Id: Ibd3c1c4dd98b297c8268a4bfed984addf96bce2f
Fixes Bug:1534464