VPNaaS: modify the models to allow multiple local cidrs
Previously, the local cidr was embeded in the vpn service object
preventing a ipsec container to handle multiple local to peer 
connections (allowed by the underlying ipsec implementation - openswan).

This patch changes the internal VPNaaS models to allow multiple local
cidrs to connect to multiple peer cidrs in a single vpn service. This
change has an impact on the VPN API (and thus anyone talking to it) as 
now the subnet field in the vpn service object has been moved to a list
of subnet cidrs in the ipsec site connections.

Change-Id: I5626a6cfeb8b29a88afaf759e8aef297c108cd67
Signed-off-by: Xavi León <xavi.leon@midokura.com>