VPNaaS: allow multiple vpn services per router
Right now, we have a single midonet vpnservice and a single
container per router. Thus, vpnservice and container objects where bound
1-1. Moreover, new ipsec site connections would hold a reference to its
sibling vpnservice object (as it comes from neutron).
It seems natural to match the upstream neutron model in which the
vpnservice is just a placeholder for a list of ipsec site connections,
and decouple the container object from vpnservice.

Change-Id: Idef93447d98ac7c901f29ec7ec214eb7fc8cffc6
Signed-off-by: Xavi León <xavi.leon@midokura.com>