Get metadata service route gateway from Dhcp
When creating RouterInterface, get next hop gateway for metadata
service route from the DHCP's server_address field, rather than
scanning all of the network's ports for a DHCP port.

In addition to not scaling well, the old approach fails when the
network has a Midonet-only port, as we try to fetch the corresponding
Neutron port, which doesn't exist for Midonet-only ports.

Change-Id: Ib3d3a2d228d7a1204f3db741c88ae40ce2b39a94
Signed-off-by: Brandon Berg <bberg@midokura.com>
Fixes: 1533629