Update patch set 1
Patch Set 1:

RFC (maybe for another patch) I noticed that the externalIp in the vpnservice proto is not necessary at all (as this info is used on the ipsec container handler to get the mac of the matching port). Maybe we could remove it and watch the neutronrouter (where the externalIp is defined) on the container handler?

Patch-set: 1
Label: Code-Review=0