Don't update gateway routes on RIF create
RouterInterfaceTranslator.translateCreate() was updating all gateway
routes for its subnet with the interface port's IP address. This is
unnecessary in the typical case (network has one router interface, in
which case it gets the gateway iP by default), and wrong in the
atpyical case (network has multiple router interfaces, in which case
we shouldn't change the gateway every time we create a new interface).

Change-Id: I115b1753bdcf7565835970a615c0cb7da1b46443
Signed-off-by: Brandon Berg <bberg@midokura.com>