Agent: fix general purpose curator to root path
A previous patch that introduced the general purpose curator resulted in
nat allocations being stored at /nat instead of $root/nat.  This patch
corrects that.

We also remove the removal of the /nat subtree inside the routers upon
router deletion.  This was added recently in order to alleviate work on
the Recycler, but is no longer necessary as we've added improvements for
NAT reservations and recycling in previous patches.

Change-Id: I571462e6bcd2b53ad021de5b8dc0bfa10c5e38a0
Signed-off-by: Galo Navarro <galo@midokura.com>