VPNaaS: support updates to the external gateway ip
Right now, ipsec container configures the ipsec connections
based on the current external ip stored in the vpnservice
(set by the vpnservice translator). We currently not watch
for updates on this external ip, so changing it would break
existing connections.

This patch adds support for watching updates on the neutron router
associated to the vpn as well as changes on the external gateway ip
associated to the gateway port of the neutron router. This patch also
cleans up unnecessary fields on the vpn service object, as now the
external gw ip is read directly from the neutron router.

Change-Id: Ib61d721933a7a03fd984d81390c6c50e35b842ab
Signed-off-by: Xavi León <xavi.leon@midokura.com>