Agent: Detect duplicate flows
This patch keeps track of ManagedFlows by index, which is used as the
sequence number in netlink flow create commands. If we try to install
a duplicate flow, we extract the sequence number from the netlink
error which we use to retrieve the flow and remove it from the
userspace data structures.

Change-Id: I02ddf617656f6e3ecf49d9d1971ffb73c1ed3ed8
Signed-off-by: Duarte Nunes <duarte@midokura.com>