Fix NPE in FlowRecording
Flow history records the UUIDs of all rules traversed. However, trace
rules are generated directly from the TraceRequestsChainMapper, and
these were not getting assigned UUIDs, so when flow history tried to
record them, it got a NullPointerException.

This change fixes that scenario, by manually setting the id of
TraceRules created by the mapper.

Change-Id: Ib8867a1b9d1b081d189c08e9f7e2e63e06865df2
Signed-off-by: Ivan Kelly <ivan@midokura.com>