Fix race in PacketWorkflow actor construction
PacketsEntryPoint was creating worker by encapsulating the call to
newPacketWorkflow and multiple calls to FlowStateTable.addShard inside a
closure, which was then called on akka worker threads. addShard is not
thread safe, so a race could occur where one shard couldn't read entries
in other shards due to having the same workerId (which is generated
non-atomically).

The fix is to move the construction of these objects outside of the
closure, so that they will be called in series.

Change-Id: I09bc236121848bc536978e8c478077b066c46367
Signed-off-by: Ivan Kelly <ivan@midokura.com>