Keystone: Add support for V3
The Keystone client implements the Keystone v3 authentication and
authorization API. This patch also modifies the cluster configuration
to allow setting the version and domain. The default Keystone
version is 3.

Currently, the clients authentication with Keystone via the REST
API must use a project (tenant) that belongs within the scope of
the domain configured in the cluster configuration. The REST API
does not provide the client the capability of specifying their
own domain.

The validation of tokens is done with an administrative account,
which can be specified, similarly to Keystone v2, as an admin
token or as credentials for an administrative user. When present,
the admin token takes precedence over the password credentials.

The patch does not provide the capability to list the projects,
domain or users, this will be added in a subsequent patch. Also,
the patch includes only integration tests for Keystone. Unit tests
and necessary changes to the mock Keystone will be added in a
subsequent patch.

Change-Id: Ia84dde9161a07892fe0548822b776e757e763f5f
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>