Update patch set 3
Patch Set 3: Code-Review+2

(1 comment)

The patch seems fine, given that some fixes are done later on. I have only one question though: So we dropped the reference to the neutron subnet (i.e. the local subnet) from the VpnService object and added the local subnet(s) as a list to the site connection, is that right? Was this done to ensure that an IPSecSiteConnection can be used for multiple VPN services?

Patch-set: 3
Reviewer: Alex Bikfalvi <1002751@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2