Update patch set 3
Patch Set 3: -Code-Review

Not exactly. This change was done to ensure that a single ipsec site connection can have multiple local and peer cidrs to connect to (which is actually more in line with the current vpnaas implementation in neutron/master). 

Right now, if you wanted to connect to several remote cidrs from several local cidrs, you'd be forced to create several vpn service objects for it (following the kilo/already outdated model).

Patch-set: 3
Label: Code-Review=0