A 2nd curator instance with small session timeout
This patch introduces a 2nd curator instance with a small session
timeout in the midonet backend. The small timeout allows to perform
quick failure detection of hosts that store ephemeral nodes.

Follow-up patches will introduce a new state key type that is
manipulated with this 2nd curator instance. The final goal is to
provide fast failure detection of container services using this
new key type.

Change-Id: Idb7a3bfc15e6442359b20e6bbc8a2f72e6f609dc
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>