Update patch set 1
Patch Set 1:

Nicolas: the REST API provides only a read-only interface for Keystone, right now the CLI and the MN manager are using it to list the tenants. We don't allow changes to the tenants or users (although we could if needed). Right now the only way to modify the identity credentials is from OpenStack.

Patch-set: 1
Label: Code-Review=0