Introduce a file tailer
This file tailer reads the content, line by line, from a text file
that is continuously modified. The reader also works with FIFO files
(pipes). The class implements the functionality of `tail -f` UNIX
command.

The current class distinguishes itself from the Apache implementation
in that it allows the use of a scheduling executor service, which
can be shared with other tailers or for other operations. The tailer
yields the executor thread whenever the current end of the file is
reached, or when the trailer has been reading continuously for a
certain configurable duration.

Change-Id: I7503aac5262bae6345b4c3e8f60540bdbb65b744
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>