Agent: Decrease sticky DNAT expiration
This patch significantly decreases the expiration time for DNAT,
which was too high for our current flow state implementation.

It also changes the TimedExpirationMap so that it can handle keys with
different expiration times, so that a key with a given expiration time
at the head of the queue doesn't prevent keys with smaller expirations
behind it from being cleared.

Change-Id: I26919f7f3f70d6a67f5f60b0864d370fa5c603ff
Signed-off-by: Duarte Nunes <duarte@midokura.com>