Introduce an NSDB recycler
Introduce an initial version of the NSDB recycler designed to
cleanup unused namespaces from the state storage and the state
of deleted objects.

The recycler works by scheduling periodically a recycling task that
performs the following:

(a) Verifies that the NSDB has not been recycled within a specified
    interval before the current moment. The recycling timestamp is
    stored as part of the NSDB version key (e.g. /midonet/zoom/0)
(b) It collects the current hosts.
(c) It collects the current namespaces and deletes all namespaces
    that have a creation timestamp previous to the moment of
    starting the reycling task, and do not have a coresponding host
    object.
(d) It collects the current objects for all classes.
(e) It collects the current object state for all classes, and
    deletes that state node for those objects that do not exists,
    and have a creation data previous to the moment of starting
    the recycling task.

Ref: MI-252

Change-Id: Id223240e4df940395489ee153a872602f9a75d9c
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>