REST API allows wildcard Accept headers
This modifies the API JSON provider to process wildcard Accept
headers. To this end, we add the "application/json" media type
to the list of produced media types, such that we default to
this media type for any less specific media types such as "*/*"
or "application/*". Second, we override the matching media type
method to ensure we handle wildcard media types.

Change-Id: I8c88a55151f6eab7c8e2eda2a03cbd6164c41790
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>