VPNaaS: support updates to the external gateway IP on translators
Right now, ipsec container configures the ipsec connections based on the
current external IP stored in the vpnservice (set by the vpnservice
translator). We currently not watch for updates on this external ip,
so changing it would break existing connections.

This patch adds support for modifying the external ip stored on the vpn
service objects whenever the IP of the gateway port of a VPN-enabled
router is updated. It will update the local redirect rules and vpn
service objects associated according to the new IP, which in turn will
produce an event on the agent so the ipsec connections can be
re-stablished.

Change-Id: Ia2804554c5f6cd4f96959ec3a963696d12671eae
Signed-off-by: Xavi León <xavi.leon@midokura.com>