Remove dependency on jmap in watchdog
The watchdog depends on jmap to take a heap dump when the agent becomes
unresponsive. This is bad, because jmap doesn't come with the JRE, but
only in the JDK. Midolman packages only depend on the JRE, not the JDK.

To get around this, this patch takes a coredump instead of a heap
dump. This can then be turned into a hprof dump using jmap on another
machine.

Change-Id: I4f92656fc4c47ce6e33b4aeb49f1a9b1dc332c6a
Signed-off-by: Ivan Kelly <ivan@midokura.com>