Make sure we start fail fast curator connection
When starting the midonet backend, we didn't start the fail fast curator
connection. Moreover, the container service waits for this FFC
connection to be active before actually notifying the service start.
This made the container service to wait indefinitely (no events
received) for a connection that never was started, blocking midolman and
wdog killing it afterwards. This patch makes sure that the FFC
connections is started and reduces the timeout to wait for notification
to the default curator session timeout.

Change-Id: I915a2d810a60338ab3ff0a0a3fe7b2034a868e02
Signed-off-by: Xavi León <xavi.leon@midokura.com>