Update patch set 19
Patch Set 19:

> there's no bullet-proof solution that doesn't entail getting an ack
 > from the kernel when you create a flow. so, here we just circle
 > through a 28-bit curIndex value and assume that we won't reuse a
 > value (2^18 flow creates away) before processing a duplicate flow
 > request.

A fool proof solution would be to check if the flow for that match has been installed already. The only reason not to do it like that is GC. Memory footprint would be roughly the same as this change.

I really don't like the fact that this can leak flows. With invalidations and conntrack, I don't think it's infeasible that the index could loop around when there's heavy load and start hitting the scenario where it loses flows. This would be a pain to debug.

Patch-set: 19
Label: Code-Review=0