Update patch set 19
Patch Set 19:

that's exactly what i said: get an ack from the kernel for the netlink operation. which  adds too much latency and is an extra burden on the netlink channel (we did it in the past). i really, really don't want to do down this path again.

i don't understand why invalidations and conntrack are relevant. 

to cycle back to a used index we have to install 2^28 flows with a single simulation thread; even installing 1M flows per second, we'd have about 5 minutes to process a duplicate message, which i think is feasible. also, remember that with the current ovs kernel module, we can install at most 80k flows per second. if this is still a worry, i can retrieve my log2 code and ensure that with 1 simulation thread we have 32 bits available :-)

Patch-set: 19
Label: Code-Review=0