VPNaaS: depend on libreswan
openswan and libreswan differ on the ipsec commands. One option could be
to diferentiate between both in the code itself. As packages are already
available by the libreswan community (and also in our own repository) it
seems easy to just depend on it and document where to get it from.

Change-Id: I92e9dced9ae2cbf95b78e2ca2f085de2f6729c5e
Signed-off-by: Xavi León <xavi.leon@midokura.com>