Populate IPAddrGroup objects in case that it's already loaded
There is a bug where if a chain had a rule that references an
IPAddrGroup, and another rule was added that referenced the same
IPAddrGroup, then the ipAddrGroup{Src,Dst} would not be populated, since
they were only being populated when the IPAddrGroup object was loaded
for the first time. However, they were already being cached.

The fix checks if newly added rules have IPAddrGroup references, and if
they do, whether we already have the object in the cache. If not, the
reference increment will load it and it will be set as it was being set
before. If it does exist, it populated the ipAddrGroup{Src,Dst} object.

Ref MI-190

Change-Id: I9cf011ae05a8a79384c95ac14d9e24de109b4fa9
Signed-off-by: Ivan Kelly <ivan@midokura.com>