Update patch set 1
Patch Set 1:

Update to the trans. spec is of course desirable.  But I'd like to see some automated test covering for regressions. There is no reason why this can't include a change to RouterTranslatorTest. 

Ideally, I'd also like to see a simulation running with a neutron router (this is possible both in unit tests and mdts).

Patch-set: 1