MDTS: change to local repo in overrides
Until now, overrides installed packages with dpkg -i which did not
resolve package dependencies. Midonet-sandbox now provides a way
to install packages from a local repo which allows us to automatically
update the image in the CI every time a dependency is changes in one of
our packages. Additionally, we pull the current downstream neutron kilo
package to test vpnaas in the CI (will be done in the jobs in the
future).

Ref.: MI-523
Change-Id: I581080076da49e3276f5693678708fba15b9e075
Signed-off-by: Xavi León <xavi.leon@midokura.com>