Enable OVSDB tests, remove non usable ones
Re-enable OvsdbVtepDataTest cases that are passing.

Also, remove two failing scenarios:

    scenario("ucast mac updates")
    scenario("mcast mac updates")

Which manifiests itself in:

    2016.02.09 12:33:16.892 WARN  [pool-6-thread-1] vtep-59_62_50_39:6632 -  Updating remote MAC tables failed
    java.lang.IllegalArgumentException: Invalid UUID string: locator_id
	    at java.util.UUID.fromString(UUID.java:194) ~[na:1.8.0_40-ea]
	    at org.midonet.southbound.vtep.OvsdbUtil$.fromOvsdb(OvsdbUtil.scala:62) ~[main/:na]
	    at org.midonet.southbound.vtep.OvsdbUtil.fromOvsdb(OvsdbUtil.scala) ~[main/:na]
	    at org.midonet.southbound.vtep.schema.Table.extractUuid(Table.java:201) ~[main/:na]
	    at org.midonet.southbound.vtep.schema.UcastMacsTable.parseLocator(UcastMacsTable.java:65) ~[main/:na]
	    at org.midonet.southbound.vtep.schema.UcastMacsTable.parseEntry(UcastMacsTable.java:77) ~[main/:na]
	    at org.midonet.southbound.vtep.schema.UcastMacsTable.parseEntry(UcastMacsTable.java:32) ~[main/:na]
	    at org.midonet.southbound.vtep.mock.InMemoryOvsdbVtep.org$midonet$southbound$vtep$mock$InMemoryOvsdbVtep$$doOperation(MockOvsdbVtep.scala:382) ~[test/:na]

This is because the OVSDB client protocol allows inserting multiple rows to
multiple tables using a placeholder id.   For example, here we insert a row in
the Physical_Locator table and use a "locator_id" placeholder, then we insert a
row in Ucast_Mac_Remote (or Mcast_Mac_Remote) with a reference to the created
phys. locator, and use the placeholder "locator_id" to make it work. Our code
isn't able to figure this out, and fails when parsing locator_id as an id.

Removing makes sense at this point because at the time we wrote the tests we
didn't have any other means to test VTEP functionality.  Right now we do have
several MDTS test cases using VTEP emulators with real OVSDB instances, so the
cases that we're removing here are properly tested anyway.

As a long term evolution, a good idea would be to remove the OVSDB mock
altogether from the unit tests, and promote them to integration tests against a
dockerized VTEP, running all from gradle.

Ref.: MI-40
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: Ia97a3c051d44d809e54784f22422b5c421d3f6a1