DPE generates tunnel keys
Rather than relying on zookeeper to generate tunnel keys, which will
eventually lead to exhaustion, generate tunnel keys when the port
becomes active. This greatly increases the number of available tunnel
keys for use over the lifetime of a cluster.

For 1.9 the models are different than for 5.0, so this patch looks quite
different from the 5.x implementation of the same
functionality (4e4e22, 10d9f7 & e2daed).

Ref: MI-112

Change-Id: I4bf2b1f8cfdf7fc564aeab2209d4653a7d8a77f5
Signed-off-by: Ivan Kelly <ivan@midokura.com>