MDTS: remove non-necessary apt-get
This was necessary to pass the gates on the libreswan dependency before
the images in the CI were updated to reflect that a new package was
downloaded. The images are already updated so remove it to reduce the
likelyness of connectivity failures. Also failfast in case we cannot
update from the local repositories to avoid false positives (running the
tests with an already installed package) and increase the timeout to
wait for a service to be up (external repositories may be slow to
download packages).

Change-Id: I0da181bbdea81cc9a01d66123d0a578c06fb920d
Signed-off-by: Xavi León <xavi.leon@midokura.com>