MDTS: introduce source controlled tests
That's a first step to have script jobs in jenkins controlled in the
midonet source tree. Now, gates only rely on three scripts that can be
changed independently (ci-start.sh, ci-run.sh and ci-stop.sh). In this
patch, we introduce ci-run.sh that specifies the tests to run on the
mdts gate.

Change-Id: I62d50f7016e6a222c462160d194469f4665fed7c
Signed-off-by: Xavi León <xavi.leon@midokura.com>