Keystone: Support the tenants API with Keystone v3
This adds support to the Keystone client and the REST API to list
tenants, when using Keystone version 3. Although in v3, the concept
of tenant has been replaced with projects, this offers backwards
compatibility with the tenants endpoint, when the authentication
service is set to Keystone v3.

In the future, we may extend the REST API to include project and
domain endpoints that will provide additional information specific
to those objects.

Ref: MI-579

Change-Id: I8566e821cc9822c01172dbad5e0f035a3b3a962e
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>