Timeout when agent fails to stop actors
Currently, if an actor from the MidoNet agent crashed, the agent
would wait indefinitely when trying to shut down the actor system.

This change adds a timeout to this wait, and if the timeout
expires it continues shutting down the actor system dispatcher.
The actors service will stop sucessfully, ensuring that the agent
does not create a new dump.

Fixes: MI-411

Change-Id: I85e22ffdeaaab2c43f6691bd71fae75eefc86b1e
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>