VPNaaS: Container handlers expose configuration
This patch modifies the ContainerHandler interface to return the
container configuration upon successful completion of the create,
update and delete methods. This ensures that the last applied
configuration is available to the container service, which can
log it.

The configuration must be a single line string, that is customized
by each container handler.

In addition, the handler introduces a new cleanup method that
receives the container configuration, and performs the container
cleanup.

Ref: MI-335

Change-Id: Ia63b9477e48bfae582ed352c73fc38493626ec03
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>