VPNaaS: Containers expose configuration and cleanup
This patch modifies the ContainerHandler interface to return
a status observable, instead of a health observable. This
allows the container to report both the current configuration
and the health updates.

This also fixes an issue, where if the container namespace was
not created during the first setup but rather during a subsequent
update, it was not reported in the NSDB container status, because
the container service would never updated. Therefore, now the
container namespace is embedded directly in the health reports.

Container should report their configuration whenever it changes,
such that the container service can keep track of the
containers that are currently active and can peform cleanup
tasks.

The configuration must be a single line string, that is customized
by each container handler.

Currently, the container service does not use the container
configuration updates. Likewise, the IPSec container does not
send configuration reports, and it does not support cleanup.
These will be added in subsequent patches.

Ref: MI-335

Change-Id: Ia63b9477e48bfae582ed352c73fc38493626ec03
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>