Hide the observable cache keys in VT
VirtualTopology maintained its observable cache using class tag keys
(and the object ID).  RouterMapper instantiates a PortMapper observable
using the key (classTag[RouterPort], portId), while the PortMapper
internally does so using the key (classTag[Port], portId).  As a
consequence, when a router port is deleted, the PortMapper observable
does not get removed, and when a new router port with the same ID is
created, the obsolete PortMapper is fetched from the observable cache.

This patch fixes this issue by hiding the class tag to access the
observable cache inside the VirtualTopology, so that both Port and
RouterPort share the same key to access the observable cache.

This patch also includes a fix in DeviceMapper where a return statement
in call method caused NonLocalReturnControl error in the unit test.

Fixes MI-601

Change-Id: Ia3987c2025efaf1d0b2940f0a893f9e973fd7bcc