VPNaaS: Introduce a container logger
A container logger is responsible for logging the container
operations that are local to a MidoNet agent to a log directory.
This provides a record of the current containers and their
name, such that they can be removed after an agent failure.

The container log directory is written to the same directory as the
agent log, which is configurable using the system property
`midolman.log.dir`. The container log directory name is configurable
using the agent configuration.

The logger supports writing container operations to the log
directory as follows. When launching a container, it creates a
new file in the log directory, with the container name and type.
When a container is stopped, the logger deletes the corresponding
file.

Ref: MI-335

Change-Id: Ie7a8bf0440da756e63dc78f46f63ef0e61192eba
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit 6697b11d9f3cc647f139386b92a4ecd8636334b9)