VPNaaS: Introduce a container logger
A container logger is responsible for logging the container
operations that are local to a MidoNet agent to a log file.
This provides a record of the current containers and their
configurations, such that they can be removed after an agent
failure.

The container log file is written to the same directory as the
agent log, which is configurable using the system property
`midolman.log.dir`. The container log file name is configurable
using the agent configuration.

The logger supports writing container operations to the log file,
one per line. The logger also supports reading the log file, and
consolidating the currently started containers.

Ref: MI-335

Change-Id: Ie7a8bf0440da756e63dc78f46f63ef0e61192eba
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>