VPNaaS: Container service cleans existing containers
This patch integrates the container service and the container logger
and the container handler cleanup method as follows. Upon startup,
the service loads the current containers log to determine the
current containers.

If any containers are present, then the service asynchronously
enqueues a task to cleanup the contaniner, where the task is
executed on the containers thread. If the cleanup fails, the
failure is logged, but it does not prevent the service to start.

The cleanup of every container is handled independently, such that
the failure to cleanup one container does not impact the others.

Following a succesful loading of the current containers log, the
log is flushed.

Ref: MI-335

Change-Id: I59f0e6ba666f89ec6cfae988ed536ca2b51ca32c
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>